pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "powerctl"
version = "0.1.0"
authors = [
    {name = "Diego Cortassa", email = "diego.cortassa@gmail.com"},
]
description = "Remote power management for workstations and servers (iDRAC, iLO, AMT, SSH)."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
license-files = ["LICENSE.md"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]

# No mandatory dependencies — all protocol libs are optional extras.
dependencies = []

[project.optional-dependencies]
ssh = ["asyncssh>=2.14", "bcrypt>=5.0"]
winrm = ["pywinrm>=0.4"]
all = ["asyncssh>=2.14", "bcrypt>=5.0", "pywinrm>=0.4"]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "mypy>=1.9",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/diegocortassa/powerctl"
Issues = "https://github.com/diegocortassa/powerctl/issues"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]


# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "T10",    # flake8-debugger (replaces debug-statements hook)
]

[tool.ruff.lint.isort]
known-first-party = ["powerctl"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
